+++ mne/channels/__init__.py
@@ -8,4 +8,5 @@ from .layout import (Layout, make_eeg_layout, make_grid_layout, read_layout,
 from .montage import read_montage, read_dig_montage, Montage, DigMontage
 
 from .channels import (equalize_channels, rename_channels, fix_mag_coil_types,
-                       read_ch_connectivity, _get_ch_type)
+                       read_ch_connectivity, _get_ch_type,
+                       find_ch_connectivity)