+++ mne/channels/tests/test_channels.py
@@ -14,12 +14,15 @@ from scipy.io import savemat
 from numpy.testing import assert_array_equal
 from nose.tools import assert_raises, assert_true, assert_equal, assert_false
 
-from mne.channels import rename_channels, read_ch_connectivity
-from mne.channels.channels import _ch_neighbor_connectivity
-from mne.io import read_info, read_raw_fif
+from mne.channels import (rename_channels, read_ch_connectivity,
+                          find_ch_connectivity)
+from mne.channels.channels import (_ch_neighbor_connectivity,
+                                   _compute_ch_connectivity)
+from mne.io import read_info, read_raw_fif, read_raw_ctf, read_raw_bti
 from mne.io.constants import FIFF
 from mne.utils import _TempDir, run_tests_if_main
 from mne import pick_types, pick_channels
+from mne.datasets import testing
 
 base_dir = op.join(op.dirname(__file__), '..', '..', 'io', 'tests', 'data')
 raw_fname = op.join(base_dir, 'test_raw.fif')
@@ -214,5 +217,43 @@ def test_get_set_sensor_positions():
     raw1._set_channel_positions([[1, 2, 3]], [ch_name])
     assert_array_equal(raw1.info['chs'][13]['loc'],
                        raw2.info['chs'][13]['loc'])
+
+
+@testing.requires_testing_data
+def test_find_ch_connectivity():
+    """Test computing the connectivity matrix."""
+    data_path = testing.data_path()
+
+    raw = read_raw_fif(raw_fname, preload=True)
+    sizes = {'mag': 828, 'grad': 1700, 'eeg': 386}
+    nchans = {'mag': 102, 'grad': 204, 'eeg': 60}
+    for ch_type in ['mag', 'grad', 'eeg']:
+        conn, ch_names = find_ch_connectivity(raw.info, ch_type)
+        # Silly test for checking the number of neighbors.
+        assert_equal(conn.getnnz(), sizes[ch_type])
+        assert_equal(len(ch_names), nchans[ch_type])
+    assert_raises(ValueError, find_ch_connectivity, raw.info, None)
+
+    # Test computing the conn matrix with gradiometers.
+    conn, ch_names = _compute_ch_connectivity(raw.info, 'grad')
+    assert_equal(conn.getnnz(), 2680)
+
+    # Test ch_type=None.
+    raw.pick_types(meg='mag')
+    find_ch_connectivity(raw.info, None)
+
+    bti_fname = op.join(data_path, 'BTi', 'erm_HFH', 'c,rfDC')
+    bti_config_name = op.join(data_path, 'BTi', 'erm_HFH', 'config')
+    raw = read_raw_bti(bti_fname, bti_config_name, None)
+    _, ch_names = find_ch_connectivity(raw.info, 'mag')
+    assert_true('A1' in ch_names)
+
+    ctf_fname = op.join(data_path, 'CTF', 'testdata_ctf_short.ds')
+    raw = read_raw_ctf(ctf_fname)
+    _, ch_names = find_ch_connectivity(raw.info, 'mag')
+    assert_true('MLC11' in ch_names)
+
+    assert_raises(ValueError, find_ch_connectivity, raw.info, 'eog')
+
 
 run_tests_if_main()